Finished the GetFamilies class and started to do some small test. Also did timing test for using that class with pretty much same results as for GetColumns

+++ src/java/org/apache/hadoop/hbase/io/GetFamilies.java
@@ -2,7 +2,7 @@ package org.apache.hadoop.hbase.io;
 
 public class GetFamilies extends AbstractGet{
   
-  public GetFamilies(byte [] row, byte [] family, byte versions, TimeRange tr){
+  public GetFamilies(byte [] row, byte [] family, short versions, TimeRange tr){
     super.row = row;
     super.versions = versions;
     super.families.add(new Family(family));
